Only delete the reply command if the bot was able to send the reply

+++ src/main.js
@@ -52,8 +52,9 @@ bot.on('messageCreate', async msg => {
     if (! utils.isStaff(msg.member)) return; // Only staff are allowed to reply
 
     if (msg.attachments.length) await attachments.saveAttachmentsInMessage(msg);
-    await thread.replyToUser(msg.member, msg.content.trim(), msg.attachments, config.alwaysReplyAnon || false);
-    msg.delete();
+
+    const replied = await thread.replyToUser(msg.member, msg.content.trim(), msg.attachments, config.alwaysReplyAnon || false);
+    if (replied) msg.delete();
   } else {
     // Otherwise just save the messages as "chat" in the logs
     thread.saveChatMessage(msg);
